jimklo: fix for missing variables when uploading to redcap
This addresses the following new behavior caused by updated pandas and pycap:

- pandas deletes series names when creating a pd.MultiIndex from an existing pd.MultiIndex.
- pycap doesn't like to work with dataframes that uses a pd.MultiIndex.

This fix checks to see if we have a pd.MultiIndex and whose Series are named and not None. If
those conditions are met, reset the index without dropping the series which results in a flat
DataFrame which the current version of pycap can handle.